% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ASSIGN-package.R
\docType{data}
\name{combat_train}
\alias{combat_train}
\title{Pathway Signature Data}
\format{Data frame with control (GFP) and cell line signature RNA-Seq data.}
\source{
Bild et al.
}
\description{
Pathway signature RNA-Seq data from akt, bad, egfr, her2, igf1r, krasgv,
krasqh, and raf pathways. This data has already been batch corrected with
reference batch ComBat. This data can be combined with a test data set to
predict pathway activity for the pathways.
}
\keyword{datasets}

