% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ASSIGN-package.R
\docType{data}
\name{excludegenes}
\alias{excludegenes}
\title{Exclude Gene List}
\format{character vector of commonly differentially expressed genes}
\source{
Bild et al.
}
\description{
Overexpression signatures may contain genes that are consistantly
differentially expressed. This list was compiled based on the GFRN gene list.
These genes appear in at least 60% of all signatures at various lengths.
}
\keyword{datasets}

