% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_functions.R
\name{gather_assign_results}
\alias{gather_assign_results}
\title{Gather the ASSIGN results in a specific directory}
\usage{
gather_assign_results(path = ".")
}
\arguments{
\item{path}{The path to the directory containing ASSIGN results. The default
is the current working directory.}
}
\value{
A data frame of ASSIGN predictions from each run in the directory
}
\description{
Gather the ASSIGN results in a specific directory
}
