% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ASSIGN-package.R
\docType{data}
\name{geneList1}
\alias{geneList1}
\title{Pathway signature gene sets}
\format{List with 5 components representing each pathway. 200 signature
genes are selected for each pathway.}
\source{
Bild et al. (2006) Oncogenic pathway signatures in human cancers as
a guide to targeted therapies. Nature, 439, 353-357.
}
\description{
Signature genes for 5 oncogenic pathways.
}
\keyword{datasets}

