% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ASSIGN-package.R
\docType{data}
\name{gfrn_geneList}
\alias{gfrn_geneList}
\title{Pathway Signature Gene Lists}
\format{
List of gene lists for akt, bad, egfr, her2, igf1r, krasgv, krasqh,
and raf
}
\source{
Bild et al.
}
\description{
Pathway signature gene lists have been optimized based on correlations of
pathway activity data and protein data. The gene lists can be used to avoid
the bayesian gene selection step of ASSIGN, which will decrease the amount of
time it takes to run ASSIGN.
}
\keyword{datasets}
