% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_functions.R
\name{merge_drop}
\alias{merge_drop}
\title{Combine two data frames}
\usage{
merge_drop(x, y, by = 0, ...)
}
\arguments{
\item{x}{The first data frame to be coerced to one.}

\item{y}{The second data frame to be coerced to one.}

\item{by}{specifications of the columns used for merging. The default is
by row names}

\item{...}{arguments to be passed to or from methods.}
}
\value{
The returned data frame is the combination of x and y, with the
rownames properly assigned.
}
\description{
Combine two data frames
}
\examples{

\dontrun{
merged.df <- merge_drop(df1,df2)
}
}
