% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_functions.R
\name{pcaplot}
\alias{pcaplot}
\title{Display a PCA Plot of the Data}
\usage{
pcaplot(mat, sub, center = TRUE, scale = TRUE, plottitle = "PCA")
}
\arguments{
\item{mat}{The data frame on which to perform pca.}

\item{sub}{The number of samples in this batch, from left to right in the
data frame}

\item{center}{a logical value indicating whether the variables should be
shifted to be zero centered. The default is TRUE}

\item{scale}{a logical value indicating whether the variables should be
scaled to have unit variance before the analysis takes place. The default
is TRUE}

\item{plottitle}{The title to display above your PCA plot. The default is
"PCA".}
}
\value{
A PCA plot is displayed
}
\description{
Display a PCA Plot of the Data
}
