% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ASSIGN-package.R
\docType{data}
\name{testData1}
\alias{testData1}
\title{Gene expression profiling from cancer patients (test dataset)}
\format{Data frame with 1000 genes/probes (rows) and 111 samples (columns)}
\source{
Bild et al. (2006) Oncogenic pathway signatures in human cancers as
a guide to targeted therapies. Nature, 439, 353-357.
}
\description{
Gene expression datasets for 111 lung cancer patient samples, including 53
cases of lung adenocarcinoma and 58 cases of lung squamous carcinoma.
}
\keyword{datasets}

