\name{trainingData1}
\alias{trainingData1}
\docType{data}
\title{Gene expression profiling from cell line preturbation experiments (training dataset)}
\description{
  Gene expression datasets for 5 oncogenic pathway preturbation experiments, including B-Catenin, E2F3, MYC, RAS, and SRC pathways.
}
\usage{data(trainingData1)}
\format{
  Data frame with 1000 genes/probes (rows) and 55 samples (columns)
  %%\describe{
  %%  \item{\code{Tag_Sequence}}{a character vector}
  %%  \item{\code{Count}}{a numeric vector}
  %%}
}
\source{
  Bild et al. (2006) Oncogenic pathway signatures in human cancers as a guide to targeted therapies. Nature, 439, 353-357.
}
\keyword{datasets}
