\name{geneList1}
\alias{geneList1}
\docType{data}
\title{Pathway signature gene sets}
\description{
  Signature genes for 5 oncogenic pathways.
}
\usage{data(geneList1)}
\format{
  List with 5 components representing each pathway. 200 signature genes are selected for each pathway.
  %%\describe{
  %%  \item{\code{Tag_Sequence}}{a character vector}
  %%  \item{\code{Count}}{a numeric vector}
  %%}
}
\source{
  Bild et al. (2006) Oncogenic pathway signatures in human cancers as a guide to targeted therapies. Nature, 439, 353-357.
}
\keyword{datasets}