\name{testData1}
\alias{testData1}
\docType{data}
\title{Gene expression profiling from cancer patients (test dataset)}
\description{
  Gene expression datasets for 111 lung cancer patient samples, including 53 cases of lung adenocarcinoma and 58 cases of lung squamous carcinoma.
}
\usage{data(testData1)}
\format{
  Data frame with 1000 genes/probes (rows) and 111 samples (columns)
  %%\describe{
  %%  \item{\code{Tag_Sequence}}{a character vector}
  %%  \item{\code{Count}}{a numeric vector}
  %%}
}
\source{
  Bild et al. (2006) Oncogenic pathway signatures in human cancers as a guide to targeted therapies. Nature, 439, 353-357.
}
\keyword{datasets}
