% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ASSIGN-package.R
\docType{data}
\name{trainingData1}
\alias{trainingData1}
\title{Gene expression profiling from cell line preturbation experiments (training
dataset)}
\format{Data frame with 1000 genes/probes (rows) and 55 samples (columns)}
\source{
Bild et al. (2006) Oncogenic pathway signatures in human cancers as
a guide to targeted therapies. Nature, 439, 353-357.
}
\description{
Gene expression datasets for 5 oncogenic pathway preturbation experiments,
including B-Catenin, E2F3, MYC, RAS, and SRC pathways.
}
\keyword{datasets}

